Rename `CQRS` to `Handlers`
