--- backend/src/Examples/ExampleApp.Examples.Services/CQRS/Booking/LocationDTOExtensions.cs
@@ -1,11 +1,0 @@
-using ExampleApp.Examples.Contracts.Booking;
-using ExampleApp.Examples.Domain.Booking;
-
-namespace ExampleApp.Examples.Services.CQRS.Booking;
-
-public static class LocationDTOExtensions
-{
-    public static LocationDTO ToDTO(this Location l) => new(l.Latitude, l.Longitude);
-
-    public static Location ToDomain(this LocationDTO l) => new(l.Latitude, l.Longitude);
-}
