--- backend/src/Examples/ExampleApp.Examples.Services/CQRS/Projects/EmployeesAssignmentsTopicKeys.cs
@@ -1,14 +1,0 @@
-using ExampleApp.Examples.Contracts.Projects;
-using LeanCode.Pipe;
-
-namespace ExampleApp.Examples.Services.CQRS.Projects;
-
-public class EmployeeAssignmentsTopicKeys
-    : BasicTopicKeys<
-        EmployeeAssignmentsTopic,
-        EmployeeAssignedToProjectAssignmentDTO,
-        EmployeeUnassignedFromProjectAssignmentDTO
-    >
-{
-    public override IEnumerable<string> Get(EmployeeAssignmentsTopic topic) => [topic.EmployeeId];
-}
