--- backend/src/Examples/ExampleApp.Examples.Services/CQRS/Projects/ProjectEmployeesAssignmentsTopicKeys.cs
@@ -1,14 +1,0 @@
-using ExampleApp.Examples.Contracts.Projects;
-using LeanCode.Pipe;
-
-namespace ExampleApp.Examples.Services.CQRS.Projects;
-
-public class ProjectEmployeesAssignmentsTopicKeys
-    : BasicTopicKeys<
-        ProjectEmployeesAssignmentsTopic,
-        EmployeeAssignedToAssignmentDTO,
-        EmployeeUnassignedFromAssignmentDTO
-    >
-{
-    public override IEnumerable<string> Get(ProjectEmployeesAssignmentsTopic topic) => [topic.ProjectId];
-}
