+++ backend/src/Examples/ExampleApp.Examples.Services/Handlers/Booking/MoneyDTOExtensions.cs
@@ -1,0 +1,11 @@
+using ExampleApp.Examples.Contracts.Booking;
+using ExampleApp.Examples.Domain.Booking;
+
+namespace ExampleApp.Examples.Services.Handlers.Booking;
+
+public static class MoneyDTOExtensions
+{
+    public static MoneyDTO ToDTO(this Money m) => new((int)(m.Value * 100), m.Currency);
+
+    public static Money ToDomain(this MoneyDTO m) => new(m.Value / 100m, m.Currency);
+}
