+++ backend/src/Examples/ExampleApp.Examples.Services/Handlers/Projects/ProjectEmployeesAssignmentsTopicKeys.cs
@@ -1,0 +1,14 @@
+using ExampleApp.Examples.Contracts.Projects;
+using LeanCode.Pipe;
+
+namespace ExampleApp.Examples.Services.Handlers.Projects;
+
+public class ProjectEmployeesAssignmentsTopicKeys
+    : BasicTopicKeys<
+        ProjectEmployeesAssignmentsTopic,
+        EmployeeAssignedToAssignmentDTO,
+        EmployeeUnassignedFromAssignmentDTO
+    >
+{
+    public override IEnumerable<string> Get(ProjectEmployeesAssignmentsTopic topic) => [topic.ProjectId];
+}
